requirements.txt
requests
json
google-cloud-storage